Move ponds gen from Map class
 #11
